% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs1.design.R
\docType{data}
\name{bs1.design}
\alias{bs1.design}
\title{Choice Design from Burgess and Street}
\format{A data frame with 90 rows and 5 columns}
\usage{
data(bs1.design)
}
\description{
A dataset containing a \code{3^2*6/5/18} choice model experimental design
chosen to minimize D-error used in an example in Burgess and Street
(1992).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item A. The first attribute in the design, consisting of three levels.
  \item B. The second attribute in the design, consisting of three levels.
  \item C. The third attribute in the design, consisting of six levels.
}
}
\references{
Burgess and Street (1992).
}
\keyword{datasets}
