% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs3.design.R
\docType{data}
\name{bs3.design}
\alias{bs3.design}
\title{Third Choice Design from Burgess and Street}
\format{A data frame with 864 rows and 10 columns}
\usage{
data(bs3.design)
}
\description{
A dataset containing a \code{4^4*2^2*36/3/288} choice model experimental design
chosen to minimize D-error used in an example in Burgess and Street
(2007).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment, values from 1 to 192.
  \item Alternative. Alternative number within a question, values from 1 to 3.
  \item A. The first attribute in the design, consisting of four levels.
  \item B. The second attribute in the design, consisting of four levels.
  \item C. The third attribute in the design, consisting of four levels.
  \item D. The fourth attribute in the design, consisting of four levels.
  \item E. The fifth attribute in the design, consisting of two levels.
  \item F. The six attribute in the design, consisting of two levels.
  \item G. The seventh attribute in the design, consisting of eight levels.
  \item H. The eighth attribute in the design, consisting of 36 levels.
}
}
\references{
Burgess and Street (2007). The Construction of Optimal Stated
Choice Experiments: Theory and Methods. John Wiley && Sons.  Appendix 8 A.1.
}
\keyword{datasets}
