% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chocolate.design.R
\docType{data}
\name{chocolate.design}
\alias{chocolate.design}
\title{Choice Model Design For a Study of Chocolate Bar Preferences}
\format{A list containing the design and attribute levels.}
\usage{
data(chocolate.design)
}
\description{
A discrete choice experimental design generated
using \code{\link{ChoiceModelDesign}} examining preferences for
chocolate bars. 403 respondents were each asked six choice questions
involving choosing between four alternatives of chocolate bars. The
question asked each respondent for the choice questions was
"Which of these 2 ounce chocolate bars would you choose to buy?".
The responses are available in \code{\link{chocolate}}. The first
150 versions of the design were generated using a random design and
the second 150 versions were generated using the efficient algorithm
in \code{\link{ChoiceModelDesign}}.
}
\details{
An object of class \code{c("list", "ChoiceModelDesign")}. See
\code{\link{ChoiceModelDesign}} for a description of the
components.
}
\keyword{datasets}
