% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{ChoiceEnsemble}
\alias{ChoiceEnsemble}
\title{Create an ensemble of FitChoice models}
\usage{
ChoiceEnsemble(models, compare.only = FALSE)
}
\arguments{
\item{models}{A \code{list} of models, all of which are of class \code{FitChoice}.}

\item{compare.only}{Logical; whether to just produce a table comparing the models or
additionally combine them to make a new ensemble model.}
}
\description{
Create an ensemble of FitChoice models
}
