% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmplevels-data.R
\docType{data}
\name{choices.jmp.levels}
\alias{choices.jmp.levels}
\title{The choices made by respondents (corresponds to the JMP design file with
levels).}
\format{A data.frame with 750 rows and 12 factor variables.}
\usage{
data(jmplevels)
}
\description{
The choices made by respondents (corresponds to the JMP design file with
levels).
}
