% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmp-data.R
\docType{data}
\name{choices.jmp}
\alias{choices.jmp}
\title{The choices made by respondents (corresponds to the JMP design file).}
\format{A data.frame with 300 rows and 8 factor variables.}
\usage{
data(jmp)
}
\description{
The choices made by respondents (corresponds to the JMP design file).
}
