% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualfile-data.R
\docType{data}
\name{choices}
\alias{choices}
\title{The choices made by respondents in the Education design Sawtooth dual file
format dataset.}
\format{A data.frame with 741 rows and 12 factor variables.}
\usage{
data(dualfile)
}
\description{
The choices made by respondents in the Education design Sawtooth dual file
format dataset.
}
