% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise.design.R
\docType{data}
\name{cruise.design}
\alias{cruise.design}
\title{Choice Model Design for the cruise ship data set from the 2016 Sawtooth competition}
\format{A list containing the design and attribute levels.}
\usage{
data(cruise.design)
}
\description{
An object of class \code{c("list")} with the following elements:
\itemize{
\item \code{design} The experimental design as a matrix
\item \code{design.with.none} Same as design
\item \code{attribute.levels} The attribute levels of the attributes in a list
}
}
\keyword{datasets}
