% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cruise.R
\docType{data}
\name{cruise}
\alias{cruise}
\title{Cruise ship data from the 2016 Sawtooth competition}
\format{A data frame with 600 rows and 31 columns}
\usage{
data(cruise)
}
\description{
\itemize{
\item \code{version} - integer; version number of the design shown to the respondent
\item \code{task1} - integer; task number of first question shown to the respondent
\item \code{task2} - integer; task number of second question
\item \code{task3} - integer; task number of third question
\item \code{task4} - integer; task number of fourth question
\item \code{task5} - integer; task number of fifth question
\item \code{task6} - integer; task number of sixth question
\item \code{task7} - integer; task number of seventh question
\item \code{task8} - integer; task number of eighth question
\item \code{task9} - integer; task number of ninth question
\item \code{task10} - integer; task number of tenth question
\item \code{task11} - integer; task number of eleventh question
\item \code{task12} - integer; task number of twelveth question
\item \code{task13} - integer; task number of thirteenth question
\item \code{task14} - integer; task number of fourteenth question
\item \code{task15} - integer; task number of fifteenth question
\item \code{choice1} - integer value from 1 to 4 indicating the respondent's choice
for the first question
\item \code{choice2} - integer value from 1 to 4 indicating the respondent's choice
for the second question
\item \code{choice3} - integer value from 1 to 4 indicating the respondent's choice
for the third question
\item \code{choice4} - integer value from 1 to 4 indicating the respondent's choice
for the fourth question
\item \code{choice5} - integer value from 1 to 4 indicating the respondent's choice
for the fifth question
\item \code{choice6} - integer value from 1 to 4 indicating the respondent's choice
for the sixth question
\item \code{choice7} - integer value from 1 to 4 indicating the respondent's choice
for the seventh question
\item \code{choice8} - integer value from 1 to 4 indicating the respondent's choice
for the eighth question
\item \code{choice9} - integer value from 1 to 4 indicating the respondent's choice
for the ninth question
\item \code{choice10} - integer value from 1 to 4 indicating the respondent's choice
for the tenth question
\item \code{choice11} - integer value from 1 to 4 indicating the respondent's choice
for the eleventh question
\item \code{choice12} - integer value from 1 to 4 indicating the respondent's choice
for the twelveth question
\item \code{choice13} - integer value from 1 to 4 indicating the respondent's choice
for the thirteenth question
\item \code{choice14} - integer value from 1 to 4 indicating the respondent's choice
for the fourteenth question
\item \code{choice15} - integer value from 1 to 4 indicating the respondent's choice
for the fifteenth question
\item \code{Cruisedbefore} - Have you ever been on a cruise vacation before
    (large cruise liner with 100s or 1000s of people on it)? Yes|No
\item \code{Howmany} - To the best of your memory, how many cruise vacations have
    you been on? 1|2|3 to 5|6 to 10|More than 10|Don't know
\item \code{Intend} - How likely is it that you will take a cruise vacation sometime
    in the next 10 years? Extremely likely|Somewhat likely|Not very likely
\item \code{Abiltravel} - Does the amount of leisure time you have give you more or
    less time than the average person in the USA to travel?
    More time to travel|About the same|Less time to travel
\item \code{Abilspend} - Would you say that you have more or less available money
    than the average person in the USA to spend on travel?
    More money to travel|About the same|Less money to travel
\item \code{Motivation_1} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Beautiful scenery)
\item \code{Motivation_2} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Warm weather)
\item \code{Motivation_3} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Visiting tropical beaches)
\item \code{Motivation_4} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Visiting historic places)
\item \code{Motivation_5} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Socializing with new people)
\item \code{Motivation_6} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Relaxation)
\item \code{Motivation_7} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (The food & beverages)
\item \code{Motivation_8} Of the items listed here, which things do you most look
    forward to in a cruise vacation? (Spa treatments)
\item \code{TopDestination} Which of the following itineraries would you MOST
    prefer for a cruise vacation?
    Mexican Riviera (sailing out of Los Angeles, CA)|
    Eastern Caribbean (sailing out of Fort Lauderdale, FL)|
    Western Caribbean (sailing out of Tampa, FL)|
    Alaska (sailing out of Seattle, WA)|
    Norway and Northern Europe (sailing out of Oslo, Norway)|
    Mediterranean (sailing out of Barcelona, Spain)
\item \code{TopCruiseLine} Which of the following cruise lines would you prefer
    to cruise with? Norwegian|Disney|Royal Caribbean|Princess|Holland America|Carnival
\item \code{TopLength} Of the options, how many days would you prefer to cruise?
    7 days|8 days|9 days|10 days|11 days
\item \code{Demos1} Which of the following best describes you?
    Employed full time|Employed part time|Unemployed and looking for work|
    Unemployed and not currently looking for work|Retired|Other
\item \code{Demos2} What is your marital status?
    Single (never married)|Married|Separated|Widowed|Divorced
\item \code{Demos3} What is the highest degree or level of education you have completed?
    Less than or some high school|High school graduate|Some college, no degree|
    Associate's degree|Bachelor's degree|Master's degree|Ph.D.
}
}
\keyword{datasets}
