% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designefficiency.R
\name{DOptimality}
\alias{DOptimality}
\title{The D-optimality (D_0, D_P or D_B) of a design.}
\usage{
DOptimality(design, prior = NULL, n.rotations = 10, seed = 123)
}
\arguments{
\item{design}{The non-encoded design with the first three columns as
Version, Question, Alternative, and the remaining columns as attribute
levels.}

\item{prior}{If NULL, the D_0 optimality is computed. If a numeric vector,
the D_P optimality is computed. If a matrix with two columns, the
D_B optimality is computed.}

\item{n.rotations}{The number of random rotations when computing
D_B optimality.}

\item{seed}{The random seed when computing D_B optimality.}
}
\value{
The D-optimality.
}
\description{
The D-optimality (D_0, D_P or D_B) of a design.
}
