% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.choices.R
\docType{data}
\name{eggs.choices}
\alias{eggs.choices}
\title{Eggs choices}
\format{A data frame containing the respondent choices for the eggs data set.}
\usage{
data(eggs.choices)
}
\description{
Eggs choices
}
\keyword{datasets}
