% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs-data.R
\docType{data}
\name{eggs.data}
\alias{eggs.data}
\title{The experiment question from the Eggs.Q project.}
\format{A data.frame with 380 rows and 200 variables.}
\usage{
data(eggs)
}
\description{
The experiment question from the Eggs.Q project.
}
