% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.design.R
\docType{data}
\name{eggs.design}
\alias{eggs.design}
\title{Eggs Design}
\format{A list containing the design and attribute levels.}
\usage{
data(eggs.design)
}
\description{
\itemize{
  \item design
  \item attribute.levels
}
}
\keyword{datasets}
