% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggs.tasks.R
\docType{data}
\name{eggs.tasks}
\alias{eggs.tasks}
\title{Eggs tasks}
\format{A data frame containing the task IDs for the respondent choices
in the eggs data set.}
\usage{
data(eggs.tasks)
}
\description{
Eggs tasks
}
\keyword{datasets}
