% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{ExtractClassParameters}
\alias{ExtractClassParameters}
\title{ExtractClassParameters}
\usage{
ExtractClassParameters(obj)
}
\arguments{
\item{obj}{A FitChoice object from running LCA.}
}
\value{
A matrix of parameters where each column corresponds to a class.
    If there is only one class, this is a vector of parameters.
}
\description{
Produces a matrix of parameters from LCA.
}
