% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractparameterstats.R
\name{ExtractParameterStats}
\alias{ExtractParameterStats}
\title{Extract Parameter Statistics from a Choice Model Fit Using Hierarchical Bayes}
\usage{
ExtractParameterStats(fit)
}
\arguments{
\item{fit}{An object of class \code{"FitChoice"} produced by a call
to \code{FitChoiceModel} with \code{algorithm = "HB-Stan"}.}
}
\value{
A matrix containing a summary of the parameter samples from
    the MCMC results stored in \code{fit}, including mean, standard
    error, effective sample size, and rhat for each mean and
    standard deviation parameter.
}
\description{
Takes a Hierarchical Bayes (Stan) output and produces sample
statistics of the mean and standard deviation parameters for the
distribution from which individual coefficients are sampled.
}
\seealso{
\code{\link[rstan]{monitor}}
}
