% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast.food.design.R
\docType{data}
\name{fast.food.design}
\alias{fast.food.design}
\title{Choice Model Design For a Fast Food Study}
\format{A list containing the design and attribute levels.}
\usage{
data(fast.food.design)
}
\description{
A partial profiles discrete choice experimental design generated
using \code{\link{ChoiceModelDesign}} examining preferences for
fast food. 403 respondents were each asked twelve choice questions
involving choosing between four fast food alternatives. The
question asked each respondent for the choice questions was
"If you were ordering food delivery on typical Saturday night where nothing
special was happening, which of these would you choose? Assume the options are identical in
all other ways." The responses are available in \code{\link{fast.food}}.
}
\details{
An object of class \code{c("list", "ChoiceModelDesign")}. See
\code{\link{ChoiceModelDesign}} for a description of the
components.
}
\keyword{datasets}
