% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{FitChoiceModel}
\alias{FitChoiceModel}
\title{FitChoiceModel}
\usage{
FitChoiceModel(experiment.data, n.classes = 1, subset = NULL,
  weights = NULL, seed = 123, tasks.left.out = 0, hb.iterations = 500,
  hb.chains = 8, hb.max.tree.depth = 10, hb.adapt.delta = 0.8,
  hb.keep.samples = FALSE, no.stanfit = FALSE)
}
\arguments{
\item{experiment.data}{A data.frame from an Experiment question}

\item{n.classes}{The number of latent classes.}

\item{subset}{An optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{An optional vector of sampling or frequency weights.}

\item{seed}{Random seed.}

\item{tasks.left.out}{Number of questions to leave out for cross-validation.}

\item{hb.iterations}{The number of iterations in Hierarchical Bayes.}

\item{hb.chains}{The number of chains in Hierarchical Bayes.}

\item{hb.max.tree.depth}{http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded}

\item{hb.adapt.delta}{http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup}

\item{hb.keep.samples}{Whether to keep the samples of all the parameters in the output.}

\item{no.stanfit}{Whether to leave out the stanfit property}
}
\description{
Fit a choice-based conjoint model using methods such as Hierarchical Bayes
}
