% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{GetParameterStatistics}
\alias{GetParameterStatistics}
\title{GetParameterStatistics}
\usage{
GetParameterStatistics(stan.fit)
}
\arguments{
\item{stan.fit}{Whether to return a function that shows
user-friendly Stan warnings.}
}
\value{
A matrix containing parameter summary statistics.
}
\description{
This function returns functions that handle Stan warnings.
}
