% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{GetStanErrorHandler}
\alias{GetStanErrorHandler}
\title{GetStanErrorHandler}
\usage{
GetStanErrorHandler()
}
\value{
A function that takes an error object.
}
\description{
This function returns a function that handles Stan errors.
}
