% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{GetStanWarningHandler}
\alias{GetStanWarningHandler}
\title{GetStanWarningHandler}
\usage{
GetStanWarningHandler(show.stan.warnings)
}
\arguments{
\item{show.stan.warnings}{Whether to return a function that shows
user-friendly Stan warnings.}
}
\value{
A function that takes a warning object.
}
\description{
This function returns a function that handles Stan warnings.
}
