% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hz.design}
\alias{hz.design}
\title{Choice Design from Huber and Zwerina
A dataset containing a choice model experimental design chosen to minimize D-error used in an example in Huber and Zwerina (1996).}
\format{
A matrix with 27 rows and 5 variables.
}
\usage{
data(hz.design)
}
\description{
Choice Design from Huber and Zwerina
A dataset containing a choice model experimental design chosen to minimize D-error used in an example in Huber and Zwerina (1996).
}
\references{
Joel Huber and Klaus Zwerina (1996). The Importance of Utility Balance in Efficient Choice Designs. Journal of MArketing Research.
}
