% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{IsRServer}
\alias{IsRServer}
\title{IsRServer}
\usage{
IsRServer()
}
\value{
TRUE if running on an R server. False otherwise.
}
\description{
This function indicates if it is being run on an R server.
}
