% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{IsTestRServer}
\alias{IsTestRServer}
\title{IsTestRServer}
\usage{
IsTestRServer()
}
\value{
TRUE if running on the test R server. False otherwise.
}
\description{
This function indicates if it is being run on the test R
server.
}
