% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kessels1.design.R
\docType{data}
\name{kessels1.design}
\alias{kessels1.design}
\title{Choice Design from Kessels et al 2011, Table II, Design 1}
\format{A data frame with 16 rows and 8 columns}
\usage{
data(kessels1.design)
}
\description{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item A1, 2 levels
  \item A2, 2 levels
  \item A3, 2 levels
  \item A4, 2 levels
  \item A5, 2 levels
  \item A6, 2 levels
}
}
\references{
Kessels et al (2011). The usefulness of Bayesian optimal designs for
discrete choice experiments.
}
\keyword{datasets}
