% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentdata.R
\name{LeftOutQuestions}
\alias{LeftOutQuestions}
\title{LeftOutQuestions}
\usage{
LeftOutQuestions(n.respondents, n.questions, n.questions.left.out, seed)
}
\arguments{
\item{n.respondents}{Number of respondents.}

\item{n.questions}{Number of questions per respondent.}

\item{n.questions.left.out}{Number of questions per respondent to leave out.}

\item{seed}{Random seed.}
}
\value{
A logical matrix (\code{n.questions} by \code{n.respondents}) indicating
which questions are left out.
}
\description{
Randomly select questions to leave out for each respondent.
}
