% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{LogLikelihoodAndBIC}
\alias{LogLikelihoodAndBIC}
\title{LogLikelihoodAndBIC}
\usage{
LogLikelihoodAndBIC(stan.fit, n.parameters, sample.size,
  n.questions.left.out)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{n.parameters}{The number of HB model parameters.}

\item{sample.size}{The sample size of the analysis.}

\item{n.questions.left.out}{The number of questions left out.}
}
\value{
A list containing the log likelihood and BIC.
}
\description{
This function extracts the log-likelihood from a stanfit object
and computes the BIC.
}
