% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designefficiency.R
\name{LogSumExp}
\alias{LogSumExp}
\title{Computes log(sum(exp(x))) in a numerically-stable way.}
\usage{
LogSumExp(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Computes log(sum(exp(x))) in a numerically-stable way.
}
