% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{MatchChainClasses}
\alias{MatchChainClasses}
\title{MatchChainClasses}
\usage{
MatchChainClasses(stan.fit, n.chains, n.classes, n.variables)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{n.chains}{The number of chains in the analysis.}

\item{n.classes}{The number of classes in the analysis.}

\item{n.variables}{The number of variables in the analysis.}
}
\value{
A stanfit object with classes reordered if matching is successful.
Otherwise a warning is thrown and the original stanfit object is returned.
}
\description{
This function attempts to match classes generated by different
chains in a stanfit object.
}
