% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{Memberships}
\alias{Memberships}
\title{Memberships}
\usage{
Memberships(fit)
}
\arguments{
\item{fit}{A FitChoice object from running LCA.}
}
\value{
A vector containing respondent class memberships.
}
\description{
Produces a vector of respondent class memberships based on
    the maximum posterior probability.
}
