% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentdata.R
\name{nonMissingRespondents}
\alias{nonMissingRespondents}
\title{Returns which respondents are considered missing based on missing data
settings.}
\usage{
nonMissingRespondents(non.missing.table, n.questions.left.out, missing,
  n.questions)
}
\arguments{
\item{non.missing.table}{A logical matrix of respondents x questions
indicating which ones are not missing.}

\item{n.questions.left.out}{The number of questions to leave out.}

\item{missing}{The missing data setting.}

\item{n.questions}{The number of questions per respondent.}
}
\description{
Returns which respondents are considered missing based on missing data
settings.
}
