% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{ParameterStatisticsInfo}
\alias{ParameterStatisticsInfo}
\title{print.FitChoice}
\usage{
ParameterStatisticsInfo(parameter.statistics, parameter.names, n.classes)
}
\arguments{
\item{parameter.statistics}{Matrix containing parameter statistics from
a summary of a stan.fit object.}

\item{parameter.names}{Names of the parameters.}

\item{n.classes}{The number of classes.}
}
\value{
A string containing information about parameter statistics.
}
\description{
Produces a string mentioning the parameters with the lowest
effective sample size and highest Rhat.
}
