% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{parameterStatisticsLCA}
\alias{parameterStatisticsLCA}
\title{parameterStatisticsLCA}
\usage{
parameterStatisticsLCA(obj)
}
\arguments{
\item{obj}{A FitChoice object from running LCA.}
}
\value{
A table of parameter statistics (coefficients, standard errors,
    t-statistics and p-values).
}
\description{
Produces a table of parameter statistics for LCA.
}
