% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partialprofilesdesign.R
\name{partialProfilesDesign}
\alias{partialProfilesDesign}
\title{Construct an Optimal Partial Profile Design for Discrete Choice Experiments}
\usage{
partialProfilesDesign(levels.per.attribute, prior = NULL,
  alternatives.per.question, n.questions, n.constant.attributes, seed = 123)
}
\arguments{
\item{levels.per.attribute}{A \emph{named} vector containing the
number of levels for each attribute with names giving the
attribute names.  See the Details.}

\item{prior}{Number vector or two-column matrix containing prior
values for the model coefficients.  The vector length or number
of rows in the matrix must correspond to the number of
attributes/attribute levels specified in
\code{levels.per.attribute}.  If \code{NULL}, the prior for the
coefficients is assumed to be identically zero.  In the matrix
case, the first column is taken as the prior mean for the
coefficients and the second is taken to be the prior variances.
If only one column is present, the prior for the coefficients
is assumed to be centered at those values.}

\item{alternatives.per.question}{Numeric value indicating the
number of profiles to show per question.}

\item{n.questions}{Numeric value specifying the total number of
questions/tasks to be performed by each respondent.}

\item{n.constant.attributes}{The number of attributes to keep constant.}

\item{seed}{Integer value specifying the random seed to use for the
algorithm.}
}
\value{
\itemize{
    \item \code{design} - A numeric matrix which contains an efficient
    design.
    \item \code{d.criterion} - Numeric value of the D_p criterion of the
    design as described in the paper.
    \item \code{const.attributes.list} - List of numeric vectors
    of indices of the constant attributes in each question.
}
}
\description{
Uses the algorithms as described in D. P. Cuervo, R. Kessels, P. Goos and
K. Sorensen (2016).
}
\examples{
\dontrun{
    attribute.levels <- list(Att1 = 1:2, Att2 = 1:2, Att3 = 1:2,
                         Att4 = 1:4, Att5 = 1:4)
result <- ChoiceModelDesign(design.algorithm = "Partial profiles",
                            attribute.levels = attribute.levels,
                            prior = NULL,
                            n.questions = 18,
                            n.versions = 1,
                            alternatives.per.question = 2,
                            n.constant.attributes = 3,
                            seed = 1)
print(result)
}
}
\references{
D. P. Cuervo, R. Kessels, P. Goos and K. Sorensen (2016).
An integrated algorithm for the optimal design of stated choice experiments
with partial profiles. Transportation Research Part B 93.
\url{http://antor.uantwerpen.be/wordpress/wp-content/papercite-data/pdf/
palhazicuervo2016integrated.pdf}
}
