% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotposteriorintervals.R
\name{PlotPosteriorIntervals}
\alias{PlotPosteriorIntervals}
\title{Plot Posterior Intervals From a Choice Model Fit With Hierarchical Bayes}
\usage{
PlotPosteriorIntervals(fit)
}
\arguments{
\item{fit}{An object of class \code{"FitChoice"} produced by a call
to \code{FitChoiceModel} with \code{algorithm = "HB-Stan"}.}
}
\value{
A \code{ggplot} object that can be further customized using the
\code{ggplot2} package.
}
\description{
Takes a Hierarchical Bayes (Stan) output and produces
posterior intervals of the mean and standard deviation parameters
for the distribution from which individual coefficients are
sampled.
}
\seealso{
\url{https://www.displayr.com/convergence-hb-maxdiff/},
\code{\link{FitChoiceModel}} for an example, \code{\link{ExtractParameterStats}},
\code{\link{TracePlots}}

Other HB diagnostics: \code{\link{ExtractParameterStats}},
  \code{\link{TracePlots}}
}
