% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{PosteriorProbsFromDensities}
\alias{PosteriorProbsFromDensities}
\title{PosteriorProbsFromDensities}
\usage{
PosteriorProbsFromDensities(log.densities, log.class.weights, n.levels,
  n.classes)
}
\arguments{
\item{log.densities}{Log densities for each respondent question.}

\item{log.class.weights}{Log of the class weights.}

\item{n.levels}{The number of levels (respondents).}

\item{n.classes}{The number of classes.}
}
\value{
A matrix containing posterior probabilities of class membership
    for each respondent.
}
\description{
Compute posterior probabilities from log densities.
}
