% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{print.FitChoice}
\alias{print.FitChoice}
\title{print.FitChoice}
\usage{
\method{print}{FitChoice}(x, ...)
}
\arguments{
\item{x}{FitMaxDiff object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a FitChoice object
}
