% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{RandomClassMemberships}
\alias{RandomClassMemberships}
\title{RandomClassMemberships}
\usage{
RandomClassMemberships(n, n.classes, seed = 123)
}
\arguments{
\item{n}{The number of individuals to be assigned classes.}

\item{n.classes}{The number of classes.}

\item{seed}{The random seed.}
}
\value{
A matrix of size n.classes x n containing 1s and 0s where 1
    indicates membership.
}
\description{
Randomly assign n individuals to classes.
}
