% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{RepeatWeights}
\alias{RepeatWeights}
\title{RepeatWeights}
\usage{
RepeatWeights(weights, respondent.indices)
}
\arguments{
\item{weights}{A vector of respondent weights.}

\item{respondent.indices}{A list of question indices for each respondent.}
}
\value{
A vector of repeated weights.
}
\description{
Repeat weights for each respondent question.
}
