% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cho.ids.R
\docType{data}
\name{respondent.ids}
\alias{respondent.ids}
\title{Respondent IDs corresponding to the .cho data file Training.cho}
\format{A numeric vector of IDs}
\usage{
data(cho.ids)
}
\description{
Respondent IDs corresponding to the .cho data file Training.cho
}
