% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{RespondentParameters}
\alias{RespondentParameters}
\title{RespondentParameters}
\usage{
RespondentParameters(object)
}
\arguments{
\item{object}{A \code{FitChoice} or \code{FitMaxDiff} object.}
}
\description{
The parameters for each respondent.
}
