% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{RespondentParametersTable}
\alias{RespondentParametersTable}
\title{\code{RespondentParametersTable}}
\usage{
RespondentParametersTable(resp.pars, title, subtitle, footer)
}
\arguments{
\item{resp.pars}{A matrix of respondent parameters}

\item{title}{Table title.}

\item{subtitle}{Table subtitle.}

\item{footer}{Table footer.}
}
\description{
Produces a formattable table with histograms of respondent parameters.
}
