% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{RespondentQuestionsToLevels}
\alias{RespondentQuestionsToLevels}
\title{RespQuestionsToLevels}
\usage{
RespondentQuestionsToLevels(ind.levels)
}
\arguments{
\item{ind.levels}{A list of question indices for each level.}
}
\value{
A vector of levels.
}
\description{
Maps respondent question indices to levels.
}
