% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{RunStanSampling}
\alias{RunStanSampling}
\title{RunStanSampling}
\usage{
RunStanSampling(stan.dat, n.iterations, n.chains, max.tree.depth, adapt.delta,
  seed, stan.model, stan.file, keep.beta, ...)
}
\arguments{
\item{stan.dat}{The data to be passed to Stan.}

\item{n.iterations}{The number of iterations in the analysis.}

\item{n.chains}{The number of chains in the analysis.}

\item{max.tree.depth}{Maximum tree depth setting. See Stan documentation.}

\item{adapt.delta}{Adapt delta setting. See Stan documentation.}

\item{seed}{Random seed.}

\item{stan.model}{Complied Stan model (if running on the R server).}

\item{stan.file}{Path to Stan file (if not running on the R server).}

\item{keep.beta}{Whether retain the beta draws in the output.}

\item{...}{Additional parameters to pass on to \code{rstan::stan} and
\code{rstan::sampling}.}
}
\value{
A stanfit object.
}
\description{
Wrapper function for \code{rstan:stan} and
\code{rstan:sampling} to run Stan HB analysis.
}
