% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawtooth-data.R
\docType{data}
\name{sawtooth.data}
\alias{sawtooth.data}
\title{The experiment question from the Sawtooth competition.}
\format{A data.frame with 600 rows and 435 variables.}
\usage{
data(sawtooth)
}
\description{
The experiment question from the Sawtooth competition.
}
