% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawtooth.R
\name{SawtoothAccuracy}
\alias{SawtoothAccuracy}
\title{SawtoothAccuracy}
\usage{
SawtoothAccuracy(design.file, respondent.file, utilities.file,
  dual.response.none = FALSE)
}
\arguments{
\item{design.file}{The csv design file generated by SawtoothDualCsv. For
out-of-sample predictions, use the file from design.out.file.}

\item{utilities.file}{The csv file of respondent utilities from Sawtooth
CBC HB.}

\item{dual.response.none}{Whether the data contains dual response none.}

\item{respondent.data}{The csv response file generated by SawtoothDualCsv. For
out-of-sample predictions, use the file from respondent.out.file.}
}
\value{
If the design does not contain dual-response none, then a vector of
    prediction accuracies of the respondents is returned. Otherwise, a list
    is returned with prediction accuracies of just the second (none)
    response and the prediction accuracies of the overall response.
}
\description{
Computes the accuracies given the design, response and
    utilities files.
}
