% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sawtooth.R
\name{SawtoothDualCsv}
\alias{SawtoothDualCsv}
\title{SawtoothDualCsv}
\usage{
SawtoothDualCsv(design, respondent.data, covariates.data = NULL,
  subset = NULL, dual.response.none = FALSE,
  dual.response.none.as.none = FALSE, design.file,
  design.out.file = NULL, respondent.file, respondent.out.file = NULL,
  covariates.file = NULL, include.choice.parameters = TRUE,
  n.questions.left.out = 0, seed = 123)
}
\arguments{
\item{design}{A choice model design object.}

\item{respondent.data}{Repondent data frame containing version, tasks,
choices.}

\item{covariates.data}{Covariates data frame to be exported to a .csv file.}

\item{subset}{A filter variable for respondents.}

\item{dual.response.none}{Whether to use the "dual response none" data,
if it is present.}

\item{dual.response.none.as.none}{Whether to treat "dual response none" data
as a "none" option.}

\item{design.file}{Name of the design csv file to export.}

\item{design.out.file}{Name of the out-of-sample design csv file to
export. This is to be used as an input into SawtoothAccuracy.}

\item{respondent.file}{Name of the response csv file to export.}

\item{respondent.out.file}{Name of the out-of-sample response csv file
to export. This is to be used as an input into SawtoothAccuracy.}

\item{covariates.file}{Name of the covariates csv file to export.}

\item{include.choice.parameters}{Whether to include choice parameters to
the design.}

\item{n.questions.left.out}{Number of questions to leave out. If questions
are left out then the "out" file names need to be supplied.}

\item{seed}{Random seed.}
}
\value{
Nothing is returned, but .csv files are written to the working
     directory.
}
\description{
Creates the design and response .csv files to be used as input
    to Sawtooth CBC HB (dual csv format). Also outputs a covariates .csv
    file if covariates are supplied. Remember to select "Code variables
    using dummy coding" in the Sawtooth settings.
}
