% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbl1.design.R
\docType{data}
\name{sbl1.design}
\alias{sbl1.design}
\title{First Choice Design from Street, Burgess, and Louviere}
\format{A data frame with 32 rows and 7 columns}
\usage{
data(sbl1.design)
}
\description{
A dataset containing a \code{4^5/2/16} choice model experimental design
chosen to minimize D-error used in an example in Street, Burgess, and Louviere
(2005).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item A. The first attribute in the design, consisting of four levels.
  \item B. The second attribute in the design, consisting of four levels.
  \item C. The third attribute in the design, consisting of four levels.
  \item D. The fourth attribute in the design, consisting of four levels.
  \item E. The fifth attribute in the design, consisting of four levels.
}
}
\references{
Street, Burgess, and Louviere (2005). Quick and Easy Choice Sets:
Constructing Optimal and Nearly Optimal Stated Choice Experiments. International
Journal of Research In Marketing, 22 (4), 459-470.
}
\keyword{datasets}
