% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processcovariatedata.R
\name{ScaleNumericCovariates}
\alias{ScaleNumericCovariates}
\title{Scales numeric covariates to have a mean of zero and a standard deviation
of 0.5 as this is optimal for logit as recommended by Gelman in}
\usage{
ScaleNumericCovariates(covariates)
}
\arguments{
\item{covariates}{A matrix of covariates.}
}
\description{
Scales numeric covariates to have a mean of zero and a standard deviation
of 0.5 as this is optimal for logit as recommended by Gelman in
}
