% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw1.design.R
\docType{data}
\name{sw1.design}
\alias{sw1.design}
\title{First Choice Design from Sandor and Wendel}
\format{A data frame with 30 rows and 7 columns}
\usage{
data(sw1.design)
}
\description{
A dataset containing a \code{3^5/2/15} choice model experimental design
chosen to minimize D-error used in an example in Sandor and Wendel
(2001).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item A1. The first attribute in the design, consisting of three levels.
  \item B1. The second attribute in the design, consisting of three levels.
  \item C1. The third attribute in the design, consisting of three levels.
  \item D1. The forth attribute in the design, consisting of three levels.
  \item E1. The fifth attribute in the design, consisting of three levels.
}
}
\references{
Sandor and Wendel (2001). Designing Conjoint Experiments Using Manager's
Prior Beliefs.
}
\keyword{datasets}
