% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw2.design.R
\docType{data}
\name{sw2.design}
\alias{sw2.design}
\title{Second Choice Design from Sandor and Wendel}
\format{A data frame with 30 rows and 6 columns}
\usage{
data(sw2.design)
}
\description{
A dataset containing a \code{3^4/2/15} choice model experimental design
chosen to minimize D-error used in an example in Sandor and Wendel
(2001).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment, values from 1 to 15.
  \item Alternative. Alternative number within a question, values from 1 to 2.
  \item A. The first attribute in the design, consisting of three levels.
  \item B. The second attribute in the design, consisting of three levels.
  \item C. The third attribute in the design, consisting of three levels.
  \item D. The forth attribute in the design, consisting of three levels.
}
}
\references{
Sandor and Wendel (2001). Designing Conjoint Experiments Using Manager's
Prior Beliefs.
}
\keyword{datasets}
