% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmplevels-data.R
\docType{data}
\name{tasks.jmp.levels}
\alias{tasks.jmp.levels}
\title{The task indices given to respondents (corresponds to the JMP design file
with levels).}
\format{A data.frame with 750 rows and 12 numeric variables.}
\usage{
data(jmplevels)
}
\description{
The task indices given to respondents (corresponds to the JMP design file
with levels).
}
