% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmp-data.R
\docType{data}
\name{tasks.jmp}
\alias{tasks.jmp}
\title{The task indices given to respondents (corresponds to the JMP design file).}
\format{A data.frame with 300 rows and 8 numeric variables.}
\usage{
data(jmp)
}
\description{
The task indices given to respondents (corresponds to the JMP design file).
}
