% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualfile-data.R
\docType{data}
\name{tasks}
\alias{tasks}
\title{The task indices given to respondents in the Education design Sawtooth dual
file format dataset.}
\format{A data.frame with 741 rows and 12 numeric variables.}
\usage{
data(dualfile)
}
\description{
The task indices given to respondents in the Education design Sawtooth dual
file format dataset.
}
