% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.cho.lines.R
\docType{data}
\name{test.cho.lines}
\alias{test.cho.lines}
\title{A character vector containing the lines from a CHO file.
This is used to test the cho.lines argument in FitChoiceModel.}
\format{A character vector with 55200 elements.}
\usage{
data(test.cho.lines)
}
\description{
A character vector containing the lines from a CHO file.
This is used to test the cho.lines argument in FitChoiceModel.
}
\keyword{datasets}
