% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.labeled.design.variables.R
\docType{data}
\name{test.labeled.design.variables}
\alias{test.labeled.design.variables}
\title{A list containing labeled variables corresponding to the columns in a design
file. This is used to test the design.variables argument in FitChoiceModel.}
\format{Contains a list with the following 10 variables (the names are not
    present in the list, but in the attributes of the variables):
\itemize{
    \item Survey: the Version column
    \item ChoiceSet: the Task column
    \item ChoiceID: the Alternative column
    \item Weight: attribute 1
    \item Organic: attribute 2
    \item Charity: attribute 3
    \item Quality: attribute 4
    \item Uniformity: attribute 5
    \item Feed: attribute 6
    \item Price: attribute 7
}}
\usage{
data(test.labeled.design.variables)
}
\description{
A list containing labeled variables corresponding to the columns in a design
file. This is used to test the design.variables argument in FitChoiceModel.
}
\keyword{datasets}
