% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Utilities}
\alias{Utilities}
\title{Utilities}
\usage{
Utilities(x, scaling = c("As is", "Min = 0", "Mean = 0")[2],
  levels.order = c("As is", "Reverse", "Increasing", "Decreasing")[3],
  attr.order = c("As is", "Increasing", "Decreasing")[1],
  attr.levels.not.ordered = NULL, weights = NULL, subset = NULL,
  exclude.attributes = NULL, output = c("Line", "Column", "Bar",
  "Data")[1], colors = "#5C9AD3", line.width = 3,
  global.font.family = "Arial", global.font.color = "#2C2C2C",
  font.units = "px", grid.color = "#E1E1E1", grid.width = 1,
  level.label.font.family = global.font.family,
  level.label.font.color = global.font.color,
  level.label.font.size = 10, level.label.wrap = TRUE,
  level.label.wrap.nchar = 20, values.minimum = NULL,
  values.maximum = NULL, values.title = "Utility",
  values.title.font.family = global.font.family,
  values.title.font.color = global.font.color,
  values.title.font.size = 12, values.zero.line.color = grid.color,
  values.zero.line.width = 1,
  values.tick.font.family = global.font.family,
  values.tick.font.color = global.font.color,
  values.tick.font.size = 10,
  attr.tick.font.family = global.font.family,
  attr.tick.font.color = global.font.color, attr.tick.font.size = 12,
  attr.tick.length = 5, hovertext.decimals = 3, margin.top = 20,
  margin.bottom = 50, margin.right = 60, margin.left = 80)
}
\arguments{
\item{x}{A \code{FitChoice} object, a matrix of numeric values, or a list containing the mean utilities 
for each attribute level (e.g. \code{list(Attribute1 = c(Level1 = 0.2, Level2 = 0.4, Level3 = 0.3), 
Attribute2 = c(Level1 = 0.2, Level2 = 0.1))}).}

\item{scaling}{Controls if the mean utility is scaled. One of "As is" (no scaling),
"Min = 0", or "Mean = 0".}

\item{levels.order}{Controls the order in which the levels of each attribute are shown.
One of "As is", "Reverse", Increasing", or "Decreasing"}

\item{attr.order}{Controls the order in which the attribute are shown.
One of "As is", Increasing", or "Decreasing" (sorting by range).}

\item{attr.levels.not.ordered}{A comma-separated list of the attributes which should not be ordered.}

\item{weights}{An optional vector of sampling weights which should be the
same length as the number of respondents in \code{x}.}

\item{subset}{An optional vector specifying a subset of respondents used to take
the mean utility. It should be the same length as the number of respondents in \code{x}.}

\item{exclude.attributes}{A comma-separated list of attributes to exclude from the plot.}

\item{output}{For "Line" (default) or "Column, "Bar" a plot showing mean (rescaled)
utilities for each attribute level will be shown. For "Data", a matrix
containing the rescaled respondent parameters will be returned}

\item{colors}{The colors of the line or columns of the chart as a hex code. This can be
a single color or a vector with length equal to the number of attributes.}

\item{line.width}{Thickness of line if \code{output} is "Line".}

\item{global.font.family}{Character; font family for all occurrences of any
font attribute for the chart unless specified individually.}

\item{global.font.color}{Global font color as a named color in character format
(e.g. "black") or an a hex code.}

\item{font.units}{One of "px" or "pt".}

\item{grid.color}{The color of the grid lines.}

\item{grid.width}{Width of the grid lines in pixels.}

\item{level.label.font.family}{Character; font family for attribute level labels.}

\item{level.label.font.color}{Font color as a named color}

\item{level.label.font.size}{Integer; Font size for attribute level labels.}

\item{level.label.wrap}{Logical; whether to wrap long attribute level labels.}

\item{level.label.wrap.nchar}{Integer; number of characters in each line
when \code{level.label.wrap} is \code{TRUE}.}

\item{values.minimum}{The lower bound of the values axis. If not specified it will be
automatically determined from the utility scores}

\item{values.maximum}{The upper bound of the values axis. If not specified it will be
automatically determined from the utility scores}

\item{values.title}{Text used as the title for the values axis.}

\item{values.title.font.family}{Character; font family for values titles.}

\item{values.title.font.color}{Font color as a named color}

\item{values.title.font.size}{Integer; Font size for values titles.}

\item{values.zero.line.color}{Color of the line at \code{utility = 0}.}

\item{values.zero.line.width}{Width of the line at \code{utility = 0}. This will
be drawn over the grid.}

\item{values.tick.font.family}{Character; font family for values ticks.}

\item{values.tick.font.color}{Font color as a named color}

\item{values.tick.font.size}{Integer; Font size for values ticks.}

\item{attr.tick.font.family}{Character; font family for attribute ticks.}

\item{attr.tick.font.color}{Font color as a named color}

\item{attr.tick.font.size}{Integer; Font size for attribute ticks.}

\item{attr.tick.length}{Length of ticks}

\item{hovertext.decimals}{Number of decimals shown in hovertext.}

\item{margin.top}{Margin between plot area and the top of the
graphic in pixels}

\item{margin.bottom}{Margin between plot area and the bottom of the
graphic in pixels}

\item{margin.right}{Margin between plot area and the right of the
graphic in pixels}

\item{margin.left}{Margin between plot area and the left of the
graphic in pixels}
}
\description{
Create plot showing mean utility or return matrix of per-respondent values
}
\examples{
xx <- list(Feed = list(Grain = 0.2, Vegetables = 0.1), 
           Weight = list(`55g` = 0.4, `60g` = 0.5, `75g` = 0.6))
Utilities(xx)
}
