% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs2.design.R
\docType{data}
\name{bs2.design}
\alias{bs2.design}
\title{Second Choice Design from Burgess and Street}
\format{A data frame with 576 rows and 11 columns}
\usage{
data(bs2.design)
}
\description{
A dataset containing a \code{4^5*2*8*8/3/192} choice model experimental design
chosen to minimize D-error used in an example in Burgess and Street
(1992).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item A1. The first attribute in the design, consisting of four levels.
  \item B1. The second attribute in the design, consisting of four levels.
  \item C1. The third attribute in the design, consisting of four levels.
  \item D1. The fourth attribute in the design, consisting of four levels.
  \item E1. The fifth attribute in the design, consisting of four levels.
  \item F1. The six attribute in the design, consisting of two levels.
  \item G1. The seventh attribute in the design, consisting of eight levels.
  \item G1. The eighth attribute in the design, consisting of eight levels.
}
}
\references{
Burgess and Street (1992).
}
\keyword{datasets}
