% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentaldesign.R
\name{ChoiceModelDesign}
\alias{ChoiceModelDesign}
\title{Choice modelling experimental design}
\usage{
ChoiceModelDesign(design.algorithm, attribute.levels, n.questions,
  alternatives.per.question, prohibitions = NULL,
  output = "Unlabelled design")
}
\arguments{
\item{design.algorithm}{The algorithm used to create the design. One of \code{"random"},
\code{"shortcut"} and \code{"enumerated"}.}

\item{attribute.levels}{\code{\link{list}} of \code{\link{vector}}s of the labels of
levels for each attribute.}

\item{n.questions}{Integer; the number of questions asked to each respondent.}

\item{alternatives.per.question}{Integer; the number of alternative products
shown in each question.}

\item{prohibitions}{\code{\link{list}} of \code{\link{vector}}s where each
\code{\link{vector}} is a prohibited alternative consisting of the indices of each
prohibited level.}

\item{output}{TODO}
}
\value{
A list with components
\itemize{
\item \code{design} - an array of dimensions (number of questions by alternatives per
question by number of attributes) where each value is the index of a level.
\item \code{...} - other diagnostics according to \code{design.algorithm}.
}
}
\description{
Creates choice model experimental designs according to a given algorithm.
}
