% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentaldesign.R
\name{ChoiceModelDesign}
\alias{ChoiceModelDesign}
\title{Choice modelling experimental design}
\usage{
ChoiceModelDesign(design.algorithm, attribute.levels, n.questions,
  alternatives.per.question, prohibitions = NULL,
  output = "Unlabelled design")
}
\arguments{
\item{design.algorithm}{The algorithm used to create the design. One of \code{"Random"},
\code{"Shortcut"}, \code{"Balanced overlap"} and \code{"Complete enumeration"}.}

\item{attribute.levels}{\code{\link{list}} of \code{\link{vector}}s of the labels of
levels for each attribute.}

\item{n.questions}{Integer; the number of questions asked to each respondent.}

\item{alternatives.per.question}{Integer; the number of alternative products
shown in each question.}

\item{prohibitions}{Character \code{\link{matrix}} where each row is a prohibited
alternative consisting of the levels of each attribute. If a level is missing or
is \code{"All"} then all levels of that attribute in combination with the other
attribute levels are prohibited.}

\item{output}{TODO}
}
\value{
A list with components
\itemize{
\item \code{design} - an array of dimensions (number of questions by alternatives per
question by number of attributes) where each value is the index of a level.
\item \code{...} - other diagnostics according to \code{design.algorithm}.
}
}
\description{
Creates choice model experimental designs according to a given algorithm.
}
