% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{ComputeRespPars}
\alias{ComputeRespPars}
\title{ComputeRespPars}
\usage{
ComputeRespPars(stan.fit, par.names, subset, parameter.scales = NULL,
  all.names = NULL)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{par.names}{Parameter names}

\item{subset}{Subset vector}

\item{parameter.scales}{Scale factors for numeric parameters.}

\item{all.names}{All parameter names, including those set to zero (excluded
from beta) due to dummy coding.}
}
\value{
A matrix of respondent parameters
}
\description{
Compute respondent parameters from a stanfit object.
}
