% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{ComputeRespPars}
\alias{ComputeRespPars}
\title{ComputeRespPars}
\usage{
ComputeRespPars(stan.fit, var.names, subset, variable.scales = NULL)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{var.names}{Variable names}

\item{subset}{Subset vector}

\item{variable.scales}{Scale factors for numeric parameters.}
}
\value{
A matrix of respondent parameters
}
\description{
Compute respondent parameters from a stanfit object.
}
