% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateExperiment.R
\name{CreateExperiment}
\alias{CreateExperiment}
\title{Create an experimental design}
\usage{
CreateExperiment(levels.per.attribute, n.prohibitions = 0, seed = 12345)
}
\arguments{
\item{levels.per.attribute}{Numeric \code{\link{vector}} of the number of levels
per attribute.}

\item{n.prohibitions}{Integer; number of prohibitions.}

\item{seed}{Integer; random seed to be used by the algorithms.}
}
\value{
A list with components
\itemize{
\item \code{attribute.levels} - a \code{\link{list}} of \code{\link{vector}}s of the
labels of levels for each attribute. The names of the vectors are the attribute labels.
\item \code{prohibitions} - Character \code{\link{matrix}} where each row is a prohibited
alternative consisting of the levels of each attribute.
}
}
\description{
Creates attributes and levels of an experiment, possibly with random prohibitions.
This is useful for quickly creating a design without typing in lists of levels.
}
