% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/designefficiency.R
\name{DError}
\alias{DError}
\title{Compute the D-error of an unlabeled design (according to Huber and Zwerina 1996)}
\usage{
DError(design.matrix, attribute.levels, effects = TRUE, prior = NULL,
  n.rotations = 10, seed = 123)
}
\arguments{
\item{design.matrix}{is a matrix for an unlabeled choice design in the long format, meaning that
each row describes one of the alternatives in one of the choice tasks. Complete data for
each task is spread across several rows. The columns are:
- Column 1 indicates the version number for each profile}

\item{attribute.levels}{is a vector of numbers indicating how many levels are in each attribute. The order should}

\item{effects}{is a boolean parameter indicating whether or not the error should be computed based on}

\item{prior}{is a vector of prior parameters for the attribute levels. Keeping prior = NULL uses a flat prior}

\item{n.rotations}{The number of random rotations performed when computing
the Bayesian criterion/error when the prior mean and variance are supplied for
partial profiles.}

\item{seed}{Integer; random seed to be used by the algorithms.}
}
\description{
Compute the D-error of an unlabeled design (according to Huber and Zwerina 1996)
}
\references{
See https://faculty.fuqua.duke.edu/~jch8/bio/Papers/Huber%20Zwerina%201996%20Marketing%20Research.pdf
}
