% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{ExtractBetaDraws}
\alias{ExtractBetaDraws}
\title{ExtractBetaDraws}
\usage{
ExtractBetaDraws(stan.fit, max.draws = 100)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{max.draws}{Maximum draws per respondent per parameter.}
}
\value{
A 3D array of beta draws.
}
\description{
This function extracts beta draws from a stanfit object.
}
