% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{FitChoiceModel}
\alias{FitChoiceModel}
\title{FitChoiceModel}
\usage{
FitChoiceModel(experiment.data = NULL, cho.file = NULL,
  design.file = NULL, attribute.levels.file = NULL, choices = NULL,
  questions = NULL, n.classes = 1, subset = NULL, weights = NULL,
  seed = 123, tasks.left.out = 0, normal.covariance = "Full",
  hb.iterations = 500, hb.chains = 8, hb.max.tree.depth = 10,
  hb.adapt.delta = 0.8, hb.keep.samples = FALSE, hb.stanfit = TRUE,
  hb.prior.mean = 0, hb.prior.sd = 5, hb.warnings = TRUE,
  hb.beta.draws.to.keep = 0, include.choice.parameters = TRUE, ...)
}
\arguments{
\item{experiment.data}{A data.frame from an Experiment question}

\item{cho.file}{The file path to a cho file.}

\item{design.file}{The file path to a Sawtooth design file.}

\item{attribute.levels.file}{A dataframe or matrix where each column
contains the level names of an attribute.}

\item{choices}{A data.frame of choices made by respondents for each
question.}

\item{questions}{A data.frame of IDs of questions presented to the
respondents.}

\item{n.classes}{The number of latent classes.}

\item{subset}{An optional vector specifying a subset of observations to be
used in the fitting process.}

\item{weights}{An optional vector of sampling or frequency weights.}

\item{seed}{Random seed.}

\item{tasks.left.out}{Number of questions to leave out for cross-validation.}

\item{normal.covariance}{The form of the covariance matrix for Hierarchical
Bayes. Can be 'Full, 'Spherical', 'Diagonal'.}

\item{hb.iterations}{The number of iterations in Hierarchical Bayes.}

\item{hb.chains}{The number of chains in Hierarchical Bayes.}

\item{hb.max.tree.depth}{http://mc-stan.org/misc/warnings.html#maximum-treedepth-exceeded}

\item{hb.adapt.delta}{http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup}

\item{hb.keep.samples}{Whether to keep the samples of all the parameters in
the output.}

\item{hb.stanfit}{Whether to include the stanfit property.}

\item{hb.prior.mean}{The mean for the priors of the mean parameters
theta_raw. This is passed as a numeric vector with each element
corresponding to an attribute, or a scalar. If hb.prior.mean is nonzero for
a categorical attribute, the attribute is treated as ordered categorical and
hb.prior.mean controls the offsets from the base attribute.}

\item{hb.prior.sd}{The standard deviations for the priors of the mean
parameters theta_raw. This is passed as a numeric vector with each element
corresponding to an attribute, or a scalar. If hb.prior.mean is nonzero for
a categorical attribute, the attribute is treated as ordered categorical and
hb.prior.sd controls the standard deviations of the offsets from the base
attribute.}

\item{hb.warnings}{Whether to show warnings from Stan.}

\item{hb.beta.draws.to.keep}{Maximum number of beta draws per respondent to
return in beta.draws.}

\item{include.choice.parameters}{Whether to include alternative-specific
parameters.}

\item{...}{Additional parameters to pass on to \code{rstan::stan} and
\code{rstan::sampling}.}
}
\value{
A list with the following components:
\itemize{
\item \code{respondent.parameters} A matrix containing the parameters of
each respondent.
\item \code{parameter.statistics} A matrix containing parameter statistics
such as effective sample size and Rhat.
\item \code{stan.fit} The stanfit object from the analysis.
\item \code{beta.draws} A 3D array containing sampling draws of beta for each
respondent.
\item \code{in.sample.accuracy} The in-sample prediction accuracy.
\item \code{out.sample.accuracy} The out-of-sample prediction accuracy.
\item \code{prediction.accuracies} A vector of prediction accuracies for
each respondent.
\item \code{algorithm} The type of algorithm used.
\item \code{n.questions.left.out} The number of questions left out for
out-of-sample testing.
\item \code{n.classes} The number of classes.
\item \code{n.respondents} The number of respondents.
\item \code{n.questions} The number of questions per respondent.
\item \code{n.choices} The number of choices per question.
\item \code{n.attributes} The number of attributes.
\item \code{n.variables} The number of variables in the analysis.
\item \code{time.taken} The time taken to run the analysis.
}
}
\description{
Fit a choice-based conjoint model using methods such as
Hierarchical Bayes
}
