% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hz.design.R
\docType{data}
\name{hz.design}
\alias{hz.design}
\title{Choice Design from Huber and Zwerina}
\format{A matrix with 27 rows and 5 columns}
\usage{
data(hz.design)
}
\description{
A dataset containing a choice model experimental design chosen to minimize D-error used in an
example in Huber and Zwerina (1996).
}
\details{
\itemize{
  \item Question. Question/Task number in the experiment
  \item Alternative. Alternative number within a question
  \item Attribute_1. The first attribute in the design, consisting of three levels.
  \item Attribute_2. The second attribute in the design, consisting of three levels.
  \item Attribute_3. The third attribute in the design, consisting of three levels.
}
}
\references{
Joel Huber and Klaus Zwerina (1996). The Importance of Utility Balance in Efficient
Choice Designs. Journal of Marketing Research.
}
\keyword{datasets}
