% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalbayes.R
\name{LogLikelihoodAndBIC}
\alias{LogLikelihoodAndBIC}
\title{LogLikelihoodAndBIC}
\usage{
LogLikelihoodAndBIC(stan.fit, n.parameters, n.respondents,
  n.questions.left.out, n.alternatives, n.respondent.questions, subset,
  has.rlh = TRUE, is.tricked.max.diff = FALSE)
}
\arguments{
\item{stan.fit}{A stanfit object.}

\item{n.parameters}{The number of HB model parameters.}

\item{n.respondents}{The number of respondents.}

\item{n.questions.left.out}{The number of questions left out.}

\item{n.alternatives}{The number of alternatives per question.}

\item{n.respondent.questions}{The total number of questions over all
respondents.}

\item{subset}{The subset used on the data.}

\item{has.rlh}{Whether RLH should be calculated.}

\item{is.tricked.max.diff}{Whether the model is tricked MaxDiff.}
}
\value{
A list containing the log likelihood and BIC.
}
\description{
This function extracts the log-likelihood from a stanfit object
and computes the BIC.
}
