% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latentclass.R
\name{parameterStatisticsLCA}
\alias{parameterStatisticsLCA}
\title{parameterStatisticsLCA}
\usage{
parameterStatisticsLCA(obj)
}
\arguments{
\item{obj}{A FitChoice object from running LCA.}
}
\value{
A matrix of size n.classes x n containing 1s and 0s where 1
    indicates membership.
}
\description{
Produces a table of parameter statistics for LCA.
}
