% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{RespondentParameters}
\alias{RespondentParameters}
\title{RespondentParameters}
\usage{
RespondentParameters(fit)
}
\arguments{
\item{fit}{A \code{FitChoice}, \code{FitMaxDiff},
\code{ChoiceEnsemble} or \code{MaxDiffEnsemble} object.}
}
\description{
The parameters for each respondent.
}
