% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicemodel.R
\name{RespondentParametersTable}
\alias{RespondentParametersTable}
\title{RespondentParametersTable}
\usage{
RespondentParametersTable(resp.pars, class.memberships = NULL,
  class.sizes = NULL, attribute.levels = NULL, title, subtitle, footer)
}
\arguments{
\item{resp.pars}{A matrix of respondent parameters}

\item{class.memberships}{A vector of respondent class memberships}

\item{class.sizes}{A vector of class proportions}

\item{attribute.levels}{Attribute levels as a named list of character vectors}

\item{title}{Table title.}

\item{subtitle}{Table subtitle.}

\item{footer}{Table footer.}
}
\description{
Produces a formattable table with histograms of respondent parameters.
}
